requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.0
python-multipart>=0.0.6
redis>=5.0.1
httpx>=0.26.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
python-dotenv>=1.0.0
google-genai>=1.0.0
email-validator>=2.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langgraph>=0.2.0
chromadb>=0.5.0
flashrank>=0.2.0
